bios: move I2C from liblitedram to libbase
