+++ litex/soc/software/bios/Makefile
@@ -17,6 +17,7 @@ OBJECTS = isr.o			\
 	  cmd_bios.o		\
 	  cmd_mem.o			\
 	  cmd_boot.o		\
+	  cmd_i2c.o			\
 	  cmd_spiflash.o	\
 	  cmd_litedram.o	\
 	  cmd_liteeth.o		\
