+++ litex/soc/software/include/base/i2c.h
@@ -1,8 +1,11 @@
 #ifndef __I2C_H
 #define __I2C_H
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 #include <stdbool.h>
-#include <generated/csr.h>
 
 /* I2C frequency defaults to a safe value in range 10-100 kHz to be compatible with SMBus */
 #ifndef I2C_FREQ_HZ
@@ -15,5 +18,9 @@
 void i2c_reset(void);
 bool i2c_write(unsigned char slave_addr, unsigned char addr, const unsigned char *data, unsigned int len);
 bool i2c_read(unsigned char slave_addr, unsigned char addr, unsigned char *data, unsigned int len, bool send_stop);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif /* __I2C_H */
