+++ litex/soc/software/libbase/Makefile
@@ -1,8 +1,8 @@
 include ../include/generated/variables.mak
 include $(SOC_DIRECTORY)/software/common.mak
 
-OBJECTS=exception.o libc.o errno.o crc16.o crc32.o console.o \
-	system.o id.o uart.o time.o qsort.o strtod.o spiflash.o strcasecmp.o
+OBJECTS = exception.o libc.o errno.o crc16.o crc32.o console.o \
+	system.o id.o uart.o time.o qsort.o strtod.o spiflash.o strcasecmp.o i2c.o
 
 all: crt0-ctr.o crt0-xip.o libbase.a libbase-nofloat.a
