+++ litex/soc/software/liblitedram/sdram.h
@@ -2,7 +2,6 @@
 #define __SDRAM_H
 
 #include <generated/csr.h>
-#include "i2c.h"
 
 void sdrsw(void);
 void sdrhw(void);
